+++ chef_solo/metadata.rb
@@ -1,1 +1,7 @@
 name                    'HelloWorldExample'
+maintainer       'Quaxel'
+maintainer_email 'YOUR_EMAIL'
+license          'MIT'
+description      'Container for Project cookbooks'
+long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
+version          '0.1.0'