+++ chef_solo/site-cookbooks/HelloWorld/CHANGELOG.md
@@ -5,7 +5,7 @@ This file is used to list changes made in each version of the HelloWorld cookboo
 
 0.1.0
 -----
-- [your_name] - Initial release of HelloWorld
+- [Quaxel] - Initial release of HelloWorld wich echoes 'Hello world'
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.